requirements.txt
aiogram
python-dotenv
aiosqlite
newspaper3k
requests
deep_translator
lxml_html_clean
google-generativeai
